Le player peut déplacer le personnage.
